fix: правки в работе
