--- src/utils/typeLifehooks.ts
@@ -1,7 +1,0 @@
-export enum LifeHook {
-  onMounted = "onMounted",
-  onBeforeMounted = "onBeforeMounted",
-  onCreated = "onCreated",
-  onBeforeUnmounted = "onBeforeUnmounted",
-  onUnmounted = "onUnmounted",
-}
